+++ apps/web/CHANGELOG.md
@@ -1,5 +1,12 @@
 # web
 
+## 0.1.9
+
+### Patch Changes
+
+- Updated dependencies
+  - poply@3.2.0
+
 ## 0.1.8
 
 ### Patch Changes